pyproject.toml
[project]
name = "copaw"
dynamic = ["version"]
description = "CoPaw is a **personal assistant** that runs in your own environment. It talks to you over multiple channels (DingTalk, Feishu, QQ, Discord, iMessage, etc.) and runs scheduled tasks according to your configuration. **What it can do is driven by Skills — the possibilities are open-ended.** Built-in skills include cron, PDF/Office handling, news digest, file reading, and more; you can add custom skills. All data and tasks run on your machine; no third-party hosting."
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    "agentscope==1.0.18",
    "agentscope-runtime==1.1.3",
    "httpx>=0.27.0",
    "packaging>=24.0",
    "discord-py>=2.3",
    "dingtalk-stream>=0.24.3",
    "uvicorn>=0.40.0",
    "apscheduler>=3.11.2,<4",
    "playwright>=1.49.0",
    "questionary>=2.1.1",
    "mss>=9.0.0",
    "reme-ai==0.3.1.8",
    "transformers>=4.30.0",
    "python-dotenv>=1.0.0",
    "python-socks>=2.5.3",
    "onnxruntime<1.24",
    "lark-oapi>=1.5.3",
    "python-telegram-bot>=20.0",
    "twilio>=9.10.2",
    "pywebview>=4.0",
    "aiofiles>=24.1.0",
    "paho-mqtt>=2.0.0",
    "wecom-aibot-python-sdk==1.0.2",
    "matrix-nio>=0.24.0",
    "shortuuid>=1.0.0",
    "google-genai>=1.67.0",
    "tzdata>=2024.1",
    "cryptography>=43.0.0",
    "keyring>=25.0.0",
    "pyyaml>=6.0",
    "json-repair>=0.30.0",
    "segno>=1.6.6",
    "modelscope>=1.35.0",
    "huggingface_hub>=0.20.0",
    "pillow>=10.0.0",
    # anyio 4.13.0: _deliver_cancellation busy-loop / getpid storm (CoPaw#2632)
    "anyio>=4.0.0,<4.13.0",
]

[tool.setuptools.dynamic]
version = {attr = "copaw.__version__.__version__"}

[tool.setuptools]
packages = { find = { where = ["src"] } }
include-package-data = true

[tool.setuptools.package-data]
# Dockerfile copies console build output to src/copaw/console/ (no dist subdir).
"copaw" = [
    "console/**",
    "agents/md_files/**",
    "agents/skills/**",
    "tokenizer/**",
    "security/tool_guard/rules/**",
    "security/skill_scanner/rules/**",
    "security/skill_scanner/data/**",
]

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project.scripts]
copaw = "copaw.cli.main:cli"

[project.optional-dependencies]
dev = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.23.0",
    "pre-commit>=4.2.0",
    "pytest-cov>=6.2.1",
    "hypothesis>=6.0.0",
]
local = [
    "huggingface_hub>=0.20.0",
]
llamacpp = [
    "copaw[local]",
    "llama-cpp-python>=0.3.0",
]
mlx = [
    "copaw[local]",
    "mlx-lm>=0.10.0; sys_platform == 'darwin'",
]
ollama = [
    "ollama>=0.6.1",
]
whisper = [
    "openai-whisper>=20231117",
]
full = [
    "copaw[local,ollama,llamacpp,whisper]",
    "mlx-lm>=0.10.0; sys_platform == 'darwin'",
]
enterprise = [
    # PostgreSQL async driver + ORM + migrations
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "pgvector>=0.2",
    # Redis
    "redis[hiredis]>=5.0.0",
    # Scheduler
    "croniter>=1.4",
    # Data processing
    "numpy>=1.24",
    #"agent-ui[vue]>=0.1.0",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "cryptography>=41.0.0",
    "authlib>=1.3.0",
    "itsdangerous>=2.1.2",
    "prometheus-fastapi-instrumentator>=6.1.0",
    # Object storage backends
    "aioboto3>=12.0",
    "miniopy-async>=1.0",
    "oss2>=2.18",
    "asyncssh>=2.14",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]
